add lint rule for cython compatible
